+++ gazebo/gui/LogPlayWidgetPrivate.hh
@@ -40,13 +40,19 @@ namespace gazebo
       public: bool paused;
 
       /// \brief Paused state of the simulation.
+      public: TimePanel *timePanel;
+
+      /// \brief Log start time.
+      public: common::Time startTime;
+
+      /// \brief Log end time.
+      public: common::Time endTime;
+
+      /// \brief If log is less than 1 hour long.
       public: bool lessThan1h;
 
-      /// TODO
+      /// \brief View which containes the timeline.
       public: LogPlayView *view;
-
-      /// \brief Paused state of the simulation.
-      public: TimePanel *timePanel;
 
       /// \brief Node used for communication.
       public: transport::NodePtr node;
@@ -66,24 +72,27 @@ namespace gazebo
     /// \brief Private data for the LogPlayView class
     class LogPlayViewPrivate
     {
-      /// \brief TODO
+      /// \brief Item which indicates the current time.
       public: CurrentTimeItem *currentTimeItem;
 
-      /// \brief Used to start, stop, and step simulation.
+      /// \brief Start time in milliseconds.
       public: int startTime;
 
-      /// \brief Used to start, stop, and step simulation.
+      /// \brief End time in milliseconds.
       public: int endTime;
 
-      /// \brief Used to start, stop, and step simulation.
+      /// \brief Width of this view's scene.
       public: int sceneWidth;
 
-      /// \brief Used to start, stop, and step simulation.
+      /// \brief Height of this view's scene.
       public: int sceneHeight;
 
-      /// \brief Used to start, stop, and step simulation.
+      /// \brief Margin from the ends.
       public: int margin;
-    };
+
+      /// \brief Whether the timeline has already been drawn.
+      public: bool timelineDrawn = false;
+     };
   }
 }
 #endif